pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "yolov8-training-project"
version = "0.1.0"
description = "Projeto para treinamento de modelos YOLOv8 com datasets do Roboflow"
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "YOLOv8 Training Team"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "roboflow==1.1.1",
    "ultralytics==8.0.158",
    "numpy==1.24.3",
    "PyYAML==6.0",
    "pytest==7.4.0",
    "python-dotenv==1.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.3.0",
    "flake8>=6.0.0",
    "pytest-cov>=4.1.0",
    "pre-commit>=3.3.1",
]

[tool.pytest.ini_options]
addopts = "--cov=src --cov-report=term-missing"
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"

[tool.black]
line-length = 120
target-version = ["py38", "py39", "py310"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 120 